Limit CPU usage in Verilator build #210
Verilator builds previously used all available CPUs by calling Make with the -j<cpu_count>
argument. This is fine when running on your local machine but is not fine when running the
build in a professional environment where you are submitting the job to a cluster.

This update removes the -j<cpu_count> argument from the build step.  Users wishing to use
N CPUs for a build should instead pass the MAKEFLAGS environment variable when calling
fusesoc:

  export MAKEFLAGS=-jN; fusesoc ...